move BlockFromToEvent handler
